fix contacts name can't display when dial from contact bug
N/A

miui contact send a spical url to phone when dial from contact,
but htc modify CallerInfo of framework implement and process the
request, then htc don't response the request as the data is invalid.

we use miui getCallerInfo implement of CallerInfo overlay the htc implement,
as they are similar and phone is miui's.
Signed-off-by: wangwei_b <wangwei_b@xiaomi.com>

Change-Id: I0ed33e8bfad8232e983247094d5b1154455c683f